+++ PhonebookSQL/PhonebookSQL.sln
@@ -1,0 +1,16 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PhonebookSQL", "PhonebookSQL\PhonebookSQL.csproj", "{03AC5B84-B211-4566-862D-299D1690CDB6}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{03AC5B84-B211-4566-862D-299D1690CDB6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{03AC5B84-B211-4566-862D-299D1690CDB6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{03AC5B84-B211-4566-862D-299D1690CDB6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{03AC5B84-B211-4566-862D-299D1690CDB6}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
