+++ Solution1/TicTacToe/TicTacToe.ConsoleApp/TicTacToe.ConsoleApp.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net7.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <RootNamespace>TicTacToe.Console</RootNamespace>
+    </PropertyGroup>
+
+</Project>
